ej1/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
sqlparse==0.4.2
djangorestframework
django-filter
markdown
